rename embeddings module
