+++ src/dbally/embeddings/base.py
